Refatorar CollectController, ICollectService e CollectService

+++ Service/Collect/CollectService.cs
@@ -2,16 +2,26 @@ using CollectApp.Data;
 using Microsoft.EntityFrameworkCore;
 using CollectApp.Models;
 using CollectApp.ViewModels;
+using SQLitePCL;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace CollectApp.Services
 {
     public class CollectService : ICollectService
     {
         private readonly CollectAppContext _context;
+        private readonly ISupplierService _supplierService;
+        private readonly IProductService _productService;
+        private readonly IFilialService _filialService;
+        private readonly ILogger<CollectService> _logger;
 
-        public CollectService(CollectAppContext context)
+        public CollectService(CollectAppContext context, ISupplierService supplierService, IProductService productService, IFilialService filialService, ILogger<CollectService> logger)
         {
             _context = context;
+            _supplierService = supplierService;
+            _productService = productService;
+            _filialService = filialService;
+            _logger = logger;
         }
 
         public async Task<List<Collect>> GetAllCollectsListAsycn()
@@ -19,9 +29,116 @@ namespace CollectApp.Services
             return await _context.Collects.Include(c => c.Supplier).Include(c => c.Product).Include(c => c.Filial).ToListAsync();
         }
 
-        public void AddCollect(Collect collect)
+        public async Task<List<CollectListItemViewModel>> SetCollectListItemViewModel()
         {
+            List<Collect> collects = await GetAllCollectsListAsycn();
+
+            List<CollectListItemViewModel> clivm = collects.Select(c => new CollectListItemViewModel
+            {
+                Id = c.Id,
+                CreatedAt = c.CreatedAt,
+                SupplierName = c.Supplier != null ? c.Supplier.Name : "-",
+                CollectAt = c.CollectAt,
+                ProductDescription = c.Product != null ? c.Product.Description : "-",
+                Status = c.Status,
+                Volume = c.Volume,
+                Weigth = c.Weigth,
+                Filial = c.Filial != null ? c.Filial.Name : "-",
+                ChangeStatus = new ChangeStatusCollectViewModel
+                {
+                    Id = c.Id,
+                    Status = c.Status
+                }
+            }).ToList();
+
+            return clivm;
+        }
+
+        public async Task CreateCollect(CreateCollectViewModel collectCreate)
+        {
+
+            Supplier? supplier = await _supplierService.FindSupplierAsync(collectCreate.SupplierId);
+            Product? product = await _productService.FindProductAsync(collectCreate.ProductId);
+            Filial? filial = await _filialService.FindFilialAsync(collectCreate.FilialId);
+
+            if (supplier == null || product == null || filial == null)
+            {
+                return;
+            }
+
+            Collect collect = new Collect
+            {
+                SupplierId = supplier.Id,
+                Supplier = supplier,
+                CollectAt = collectCreate.CollectAt,
+                ProductId = product.Id,
+                Product = product,
+                Volume = collectCreate.Volume,
+                Weigth = collectCreate.Weight,
+                FilialId = filial.Id,
+                Filial = filial,
+            };
+
             _context.Collects.Add(collect);
+            await SaveChangesCollectsAsync();
+        }
+
+        public async Task<EditCollectViewModel> SetEditCollectViewModel(int? id)
+        {
+            Collect? collect = await FindCollectAsync(id);
+
+            if (collect == null || collect.Supplier == null || collect.Product == null || collect.Filial == null)
+            {
+                EditCollectViewModel NotFound = new EditCollectViewModel();
+                return NotFound;
+            }
+
+            EditCollectViewModel ecvm = new EditCollectViewModel
+            {
+                Id = collect.Id,
+                SupplierId = collect.SupplierId,
+                Supplier = collect.Supplier.Name,
+                CollectAt = collect.CollectAt,
+                ProductId = collect.ProductId,
+                Product = collect.Product.Description,
+                Volume = collect.Volume,
+                Weight = collect.Weigth,
+                FilialId = collect.FilialId,
+                Filial = collect.Filial.Name,
+            };
+
+            return ecvm;
+        }
+
+        public async Task EditCollect(EditCollectViewModel collectEdit)
+        {
+            Collect? collect = await FindCollectAsync(collectEdit.Id);
+
+            if (collect == null)
+            {
+                return;
+            }
+
+            Supplier? supplier = await _supplierService.FindSupplierAsync(collectEdit.SupplierId);
+            Product? product = await _productService.FindProductAsync(collectEdit.ProductId);
+            Filial? filial = await _filialService.FindFilialAsync(collectEdit.FilialId);
+
+            if (supplier == null || product == null || filial == null)
+            {
+                return;
+            }
+
+            collect.SupplierId = supplier.Id;
+            collect.Supplier = supplier;
+            collect.CollectAt = collectEdit.CollectAt;
+            collect.ProductId = product.Id;
+            collect.Product = product;
+            collect.Volume = collectEdit.Volume;
+            collect.Weigth = collectEdit.Weight;
+            collect.FilialId = filial.Id;
+            collect.Filial = filial;
+
+            await SaveChangesCollectsAsync();
         }
 
         public async Task<int> SaveChangesCollectsAsync()
@@ -38,60 +155,35 @@ namespace CollectApp.Services
                 .FirstOrDefaultAsync(c => c.Id == id);
         }
 
-        public async Task<Supplier?> FindSupplierAsync(int? id)
+        public async Task UpdateCollectStatus(ChangeStatusCollectViewModel changeStatus)
         {
-            return await _context.Suppliers.FindAsync(id);
+            Collect? collect = await FindCollectAsync(changeStatus.Id);
+
+            if (collect == null)
+            {
+                return;
+            }
+
+            if (changeStatus.Status == null)
+            {
+                return;
+            }
+
+            collect.Status = changeStatus.Status;
+            await SaveChangesCollectsAsync();
         }
 
-        public async Task<Product?> FindProductAsync(int? id)
+        public async Task DeleteCollect(int? id)
         {
-            return await _context.Products.FindAsync(id);
-        }
+            Collect? collect = await FindCollectAsync(id);
 
-        public async Task<Filial?> FindFilialAsync(int? id)
-        {
-            return await _context.Filials.FindAsync(id);
-        }
+            if (collect == null)
+            {
+                return;
+            }
 
-        public void UpdateCollectStatus(Collect collect)
-        {
-            _context.Attach(collect);
-            _context.Entry(collect).Property(c => c.Status).IsModified = true;
-        }
-
-        public void DeleteCollect(Collect collect)
-        {
             _context.Collects.Remove(collect);
-        }
-
-        public async Task<List<Product>> GetFilteredProductsAsync(string input)
-        {
-            return await _context.Products.Where(p => p.Description.Contains(input)).ToListAsync();
-        }
-
-        public async Task<List<Supplier>> GetFilteredSuppliersAsync(string input)
-        {
-            return await _context.Suppliers.Where(s => s.Name.Contains(input)).ToListAsync();
-        }
-
-        public async Task<List<Filial>> GetFilteredFilialsAsync(string input)
-        {
-            return await _context.Filials.Where(f => f.Name.Contains(input)).ToListAsync();
-        }
-
-        public async Task<List<Supplier>> GetRegisteredSuppliersAsync()
-        {
-            return await _context.Suppliers.ToListAsync();
-        }
-
-        public async Task<List<Product>> GetRegisteredProductsAsync()
-        {
-            return await _context.Products.ToListAsync();
-        }
-
-        public async Task<List<Filial>> GetRegisteredFilialsAsync()
-        {
-            return await _context.Filials.ToListAsync();
+            await SaveChangesCollectsAsync();
         }
     }
 }
